third_party/WebKit/LayoutTests/http/tests/misc/onload-write-during-xframe-deny.html

<body onload="document.write('PASS');">
<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
<iframe src="http://127.0.0.1:8000/security/XFrameOptions/resources/x-frame-options-deny-meta-tag-subframe-in-body.html">
</iframe>
</body>
